lncli: move payment commands to separate file
